% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gact.R
\name{createDB}
\alias{createDB}
\title{Create GACT database directory structure}
\usage{
createDB(version = NULL, dbdir = NULL)
}
\arguments{
\item{version}{Character; name or version of the database (e.g., "v0.1").}

\item{dbdir}{Character; parent directory path in which the version folder will be created.}
}
\value{
A list (GAlist) containing the database version, base directory,
directory paths (GAlist$dirs), and feature names.
}
\description{
Initializes the directory tree used by downloadDB() and other GACT functions.
}
