% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sets.R
\name{createFunctionalSets}
\alias{createFunctionalSets}
\title{Create functional association sets (STRING, STITCH, Reactome)}
\usage{
createFunctionalSets(
  GAlist,
  min_combined_score = 900,
  min_interactions = 5,
  overwrite = FALSE
)
}
\arguments{
\item{GAlist}{GACT database list}

\item{min_combined_score}{Minimum combined score to filter STRING/STITCH interactions}

\item{min_interactions}{Minimum number of members to keep each set}

\item{overwrite}{Logical; if TRUE, regenerate existing outputs}
}
\value{
Updated GAlist with functional sets saved in gsets directory
}
\description{
Create functional association sets (STRING, STITCH, Reactome)
}
