% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gact.R
\name{downloadDB}
\alias{downloadDB}
\title{Download database components for genomic associations}
\usage{
downloadDB(
  GAlist = NULL,
  what = NULL,
  min_combined_score = 900,
  min_interactions = 5
)
}
\arguments{
\item{GAlist}{A list object containing the structure and paths needed for the database.}

\item{what}{A character string specifying the type of data to download.
Possible values include "db", "gsets", "gsea", "gstat",
"gbayes", "gwascatalog", "ensembl", "reactome", "string",
"stitch", "pubmed", "dgi", "drugbank", "atc", and others.}

\item{min_combined_score}{Minimum combined score for filtering (applicable if relevant).}

\item{min_interactions}{Minimum number of interactions for filtering (applicable if relevant).}
}
\value{
The updated GAlist object, including paths to the downloaded files.
}
\description{
This function facilitates downloading various components of a genomic
associations database. It allows selective downloading of different
sections such as genetic marker information, gene sets, GSEA summary
statistics, and other genomic data. The data is fetched from specified
sources and stored in a structured format.
}
\examples{
\dontrun{
  GAlist <- downloadDB(GAlist, what = "gsets")
}

}
