% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gact.R
\name{gact}
\alias{gact}
\title{Create the database Genomic Association of Complex Traits (GACT)}
\usage{
gact(
  GAlist = NULL,
  version = NULL,
  task = "download",
  dbdir = NULL,
  timeout = 3000
)
}
\arguments{
\item{GAlist}{A list object providing information and infrastructure of the gact database.}

\item{version}{The version of the gact database to use.}

\item{task}{The task to perform, either "download" or "process". By default "download".}

\item{dbdir}{The directory where the database should be stored, by default the current working directory.}
}
\value{
A list providing information and infrastructure of the gact database.
}
\description{
The \code{gact} function is designed for establishing and populating a comprehensive
database focused on genomic associations with complex traits. This function serves
two primary roles: infrastructure creation and data acquisition. It facilitates the
assembly of a structured repository encompassing single marker associations,
rigorously curated to ensure high-quality data. Beyond individual genetic markers,
the function integrates a broad spectrum of genomic entities, encompassing genes,
proteins, and an array of complexes (chemical, protein) as well as various
biological pathways. This integration aims to provide a holistic view of genomic
associations and their multifaceted relationships in the context of complex traits.
}
\examples{
\dontrun{
GAlist <- gact()
}

}
