% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gact.R
\docType{data}
\name{gact_versions}
\alias{gact_versions}
\title{Metadata describing available GACT database versions and sources}
\format{
An object of class \code{list} of length 1.
}
\usage{
gact_versions
}
\description{
This list defines all database versions available within GACT, including
species, description, creation date, and the corresponding data sources
(DOIs or URLs) for each component. These are used by \code{\link[=downloadDB]{downloadDB()}} to
resolve download targets.
}
\keyword{internal}
