% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_gact.R
\name{getDISEASES}
\alias{getDISEASES}
\title{Retrieve Disease Associations}
\usage{
getDISEASES(GAlist = NULL, ensg = NULL, ensp = NULL, what = "Integrated")
}
\arguments{
\item{GAlist}{A list object that contains directories and data required for the analysis, including disease data files and gene-protein mappings.}

\item{ensg}{A character vector of gene IDs (ENSG) to filter the disease data. If not provided, the function will use \code{ensp} to filter.}

\item{ensp}{A character vector of protein IDs (ENSP) to filter the disease data.}

\item{what}{A character string specifying the source of disease data. Options are "Integrated" (default), "Experiments", "Knowledge", or "Textmining".}
}
\value{
A data frame of disease associations filtered by the provided gene or protein IDs, sorted by the disease score in decreasing order.
}
\description{
This function retrieves disease association data based on protein or gene information from the provided GAlist object. It supports filtering by \code{ENSP} (protein) or \code{ENSG} (gene), and can fetch data from multiple sources: Integrated, Experiments, Knowledge, or Textmining datasets.
}
\examples{
# Assuming GAlist is a valid object with the necessary structure
diseases <- getDISEASES(GAlist, ensg = c("ENSG00000139618"))

# Retrieve data using protein IDs
diseases_protein <- getDISEASES(GAlist, ensp = c("ENSP00000269305"))

}
