% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_gact.R
\name{getFeatureSets}
\alias{getFeatureSets}
\title{Retrieve Feature Sets from Database}
\usage{
getFeatureSets(
  GAlist = NULL,
  feature = NULL,
  featureID = NULL,
  minsets = NULL,
  upstream = FALSE,
  downstream = FALSE,
  min_combined_score = 700,
  min_interactions = 5
)
}
\arguments{
\item{GAlist}{A list object providing information and infrastructure of the gact database.}

\item{feature}{A character string specifying the type of data set to retrieve.
Valid features include:
\itemize{
\item \code{"GO"}: Gene Ontology sets.
\item \code{"Pathways"}: Pathway sets.
\item \code{"ProteinComplexes"}: Protein complex sets.
\item \code{"ChemicalComplexes"}: Chemical complex sets.
\item \code{"DrugGenes"}: Drug-gene interaction sets.
\item \code{"DrugATCGenes"}: Drug ATC gene sets.
\item \code{"DrugComplexes"}: Drug complex sets.
\item \code{"DiseaseGenes"}: Disease-gene interaction sets.
\item \code{"DiseaseGenesEXP"}: Experimentally validated disease-gene sets.
\item \code{"DiseaseGenesKB"}: Knowledge-based disease-gene sets.
\item \code{"DiseaseGenesTM"}: Text-mined disease-gene sets.
\item \code{"DiseaseGenesEXPplus"}: Comprehensive experimentally validated disease-gene sets.
\item \code{"DiseaseGenesKBplus"}: Comprehensive knowledge-based disease-gene sets.
\item \code{"DiseaseGenesTMplus"}: Comprehensive text-mined disease-gene sets.
\item \code{"ATC1Genes"}, \code{"ATC2Genes"}, \code{"ATC3Genes"}, \code{"ATC4Genes"}: ATC classification gene sets.
\item \code{"GTEx"}, \code{"GTExV7"}, \code{"GTExV8"}: GTEx project eQTL sets.
\item \code{"GWAScatalog"}, \code{"GWAScatalogPlus"}: GWAS catalog sets.
\item \code{"String"}: STRING database protein interaction sets.
\item \code{"Stitch"}: STITCH database protein-chemical interaction sets.
}}

\item{featureID}{Optional; a character vector of specific feature IDs to select.}

\item{minsets}{Optional; a numeric value specifying the minimum number of sets to include.}

\item{upstream}{Optional; a logical value indicating whether to include upstream data.}

\item{downstream}{Optional; a logical value indicating whether to include downstream data.}

\item{min_combined_score}{Optional; a numeric value specifying the minimum combined score for inclusion.}

\item{min_interactions}{Optional; a numeric value specifying the minimum number of interactions for inclusion.}
}
\value{
A list of the specified feature sets. If \code{featureID} is not \code{NULL},
returns only the sets with IDs in \code{featureID}.
}
\description{
The \code{getFeatureSets} function retrieves sets of features from a \code{GAlist} object based on the specified \code{feature} argument.
It allows filtering of these feature sets using the \code{featureID} argument. This function is designed to handle a variety of
feature types, including those related to genomic, proteomic, or pharmacological information. The successful operation
of this function depends on the structure and content of the \code{GAlist} parameter.
}
\examples{
\dontrun{
sets <- getFeatureSets(GAlist, feature="Genes")
getFeatureSets(GAlist, feature="Genes", featureID=c("gene1", "gene2"))
}

}
