% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_gact.R
\name{getFeatureStat}
\alias{getFeatureStat}
\title{Get summary statistics from the database}
\usage{
getFeatureStat(
  GAlist = NULL,
  feature = NULL,
  featureID = NULL,
  studyID = NULL,
  trait = "t2d",
  threshold = 1,
  format = "data.frame",
  hyperlink = FALSE,
  cls = NULL
)
}
\arguments{
\item{GAlist}{A list object providing information and infrastructure of the gact database.}

\item{feature}{The type of statistics to retrieve from the database, such as "Markers", "Genes", "Proteins", etc.}

\item{featureID}{A character vector of specific features to extract from the database. If not specified, all features will be returned.}

\item{studyID}{The study ID to retrieve the results for. If not specified, results will be returned for all studies.}

\item{trait}{The trait to retrieve the results for. If not specified, results will be returned for all traits.}

\item{threshold}{A numeric p-value threshold used to filter results.}

\item{format}{The format to return the results in, either "list" (default) or "data.frame".}

\item{file}{The file name of the database of GSEA results.}
}
\value{
A list or data.frame of summary statistics for the specified feature, study ID, and trait.
}
\description{
This function retrieves summary statistics from the gact database of GSEA results,
such as marker statistics, gene statistics, protein statistics, etc.
}
\examples{

\dontrun{
# load example data
GAlist <- readRDS(file="GAlist.rds")

# get marker statistics
stat <- getFeatureStat(GAlist = GAlist, feature = "Markers", studyID="GWAS1")

# get gene statistics
stat <- getFeatureStat(GAlist = GAlist, feature = "Genes")

# get gene statistics for a specific feature
stat <- getFeatureStat(GAlist = GAlist, feature = "Genes", featureID = "TP53")

# get gene statistics filtered by a threshold
stat <- getFeatureStat(GAlist = GAlist, feature = "Genes", threshold = 0.05)
}
}
