% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_gact.R
\name{getGTX}
\alias{getGTX}
\title{Get GTEx data}
\usage{
getGTX(
  GAlist = NULL,
  version = "V8",
  rsids = NULL,
  ensg = NULL,
  tissue = NULL,
  cls = NULL,
  format = "data.frame"
)
}
\arguments{
\item{GAlist}{A list containing the directory path with GTEx data.}

\item{version}{A character string specifying the GTEx data version,
either "V8" or "V7".}

\item{rsids}{A vector of RSIDs for which data is to be retrieved.
If NULL, data for all available RSIDs is retrieved.}

\item{tissue}{A vector of tissue names for which data is to be retrieved.
Possible tissues include: "Adipose_Subcutaneous", "Adipose_Visceral_Omentum",
"Adrenal_Gland", "Artery_Aorta", "Artery_Coronary", "Artery_Tibial",
"Brain_Amygdala", "Brain_Anterior_cingulate_cortex_BA24", "Brain_Caudate_basal_ganglia",
"Brain_Cerebellar_Hemisphere", "Brain_Cerebellum", "Brain_Cortex",
"Brain_Frontal_Cortex_BA9", "Brain_Hippocampus", "Brain_Hypothalamus",
"Brain_Nucleus_accumbens_basal_ganglia", "Brain_Putamen_basal_ganglia",
"Brain_Spinal_cord_cervical_c-1", "Brain_Substantia_nigra", "Breast_Mammary_Tissue",
"Cells_Cultured_fibroblasts", "Cells_EBV-transformed_lymphocytes", "Colon_Sigmoid",
"Colon_Transverse", "Esophagus_Gastroesophageal_Junction", "Esophagus_Mucosa",
"Esophagus_Muscularis", "Heart_Atrial_Appendage", "Heart_Left_Ventricle",
"Kidney_Cortex", "Liver", "Lung", "Minor_Salivary_Gland", "Muscle_Skeletal",
"Nerve_Tibial", "Ovary", "Pancreas", "Pituitary", "Prostate",
"Skin_Not_Sun_Exposed_Suprapubic", "Skin_Sun_Exposed_Lower_leg",
"Small_Intestine_Terminal_Ileum", "Spleen", "Stomach", "Testis",
"Thyroid", "Uterus", "Vagina", "Whole_Blood".
If NULL, data for all available tissues is retrieved.}

\item{format}{A character string specifying the output format,
either "data.frame" or "list".}
}
\value{
A list or a data frame (based on the \code{format} parameter)
containing gene expression data.
}
\description{
This function retrieves gene expression data from the GTEx project,
based on the specified GTEx version, tissue type, and RSIDs (if provided).
It supports both GTEx V7 and V8 versions. GTEx downloaded from here:
https://www.gtexportal.org/home/downloads/adult-gtex/qtl
}
\examples{
GAlist <- list(dbdir = "path/to/GTEx/data")
gtx_data <- getGTX(GAlist, version = "V8", tissue = c("Liver", "Heart"))
}
