% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_gact.R
\name{getGWAS}
\alias{getGWAS}
\title{Retrieve GWAS Results}
\usage{
getGWAS(
  GAlist = NULL,
  studyID = NULL,
  what = "all",
  format = "list",
  rm.na = TRUE,
  rsids = NULL,
  cpra = NULL
)
}
\arguments{
\item{GAlist}{A list object providing information and infrastructure of the genomic association database.}

\item{studyID}{A vector of study IDs for which statistics are to be retrieved.}

\item{what}{Specifies the type of statistics to retrieve; "all" for all statistics or specific
types like "rsids", "b", "seb", "eaf", "ea", "nea", "z", "p" (default: "all").}

\item{format}{Specifies the format of the output; either "list" or "data.frame" (default: "list").}

\item{rm.na}{Boolean indicating whether to remove NA values (default: TRUE).}

\item{rsids}{An optional vector of SNP IDs to subset the marker statistics.}

\item{cpra}{An optional vector of chromosome-position-ref-alt values to map to marker IDs.}
}
\value{
Depending on the 'format' and 'what' parameters, the function returns a list or
data.frame of the requested marker statistics.
}
\description{
This function is a wrapper around \code{getMarkerStat()} and inherits its documentation. It retrieves GWAS results for a specified study based on the given parameters. Additional parameters can be passed to filter the results.
}
