% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gact.R
\name{getMarkerSetsDB}
\alias{getMarkerSetsDB}
\title{Get Marker Sets from database}
\usage{
getMarkerSetsDB(GAlist = NULL, feature = NULL, featureID = NULL, rsids = NULL)
}
\arguments{
\item{GAlist}{A list object providing information and infrastructure of the gact database.}

\item{feature}{A string specifying the type of biological feature for which the marker sets are to be retrieved. The options are: "Genes", "Entrez Genes", "Gene Symbol", "Proteins", "Pathways", "GO", "Protein Complexes", and "Chemical Complexes".}

\item{featureID}{The ID of the feature for which the marker sets are to be retrieved.}

\item{rsids}{A character vector of rsids to subset the marker sets.}
}
\value{
A list of marker sets for the specified feature and feature ID, where each set is a character vector of rsids.
}
\description{
This function retrieves marker sets based on a given feature and feature ID from a file. The feature can be one of the following: Genes, Inter Genes, Gene Symbol, Proteins, Pathways, GO, ProteinComplexes, ChemicalComplexes. The feature ID refers to the ID of the feature for which the marker sets are to be retrieved. The \code{rsids} argument is optional and refers to a list of SNP IDs for which the corresponding marker sets are to be retrieved.
}
\examples{
\dontrun{
sets <- getMarkerSetsDB(GAlist=GAlist, feature="Genes", featureID=c("ENSG00000165879", "ENSG00000012048"))
}
}
