% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_gact.R
\name{getSetsDB}
\alias{getSetsDB}
\title{Retrieve sets of features from a GAlist object}
\usage{
getSetsDB(
  GAlist = NULL,
  feature = NULL,
  featureID = NULL,
  upstream = FALSE,
  downstream = FALSE,
  min_combined_score = 700,
  min_interactions = 5
)
}
\arguments{
\item{GAlist}{A list object providing information and infrastructure of the gact database.}

\item{feature}{A character string specifying the type of feature set to be
retrieved. Possible values are "Entres Genes", "Genes", "Proteins", "Gene
Symbol", "GO", "Pathways", "ProteinComplexes", "ChemicalComplexes",
"ProteinComplexes2Genes", and "ChemicalComplexes2Genes".}

\item{featureID}{A character vector of IDs specifying the specific features to
be retrieved.}
}
\value{
A list of the specified feature sets. If \code{featureID} is not \code{NULL},
returns only the sets with IDs in \code{featureID}.
}
\description{
The \code{getSetsDB} function retrieves sets of features from a GAlist object. The
feature sets are specified by the \code{feature} argument, and the specific IDs of
the features can be filtered using the \code{featureID} argument.
}
\examples{
\dontrun{
sets <- getSetsDB(GAlist, feature="Genes")
getSetsDB(GAlist, feature="Genes", featureID=c("gene1", "gene2"))
}

}
