% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gact.R
\name{getStudiesDB}
\alias{getStudiesDB}
\title{#' @export
#'
writeStat <- function(GAlist=NULL, feature=NULL, featureID=NULL,
studyID=NULL, trait="T2D", threshold=1,
format="data.frame", file.csv=NULL, hyperlink=TRUE) {
stat <- getStat(GAlist=GAlist, feature=feature, featureID=featureID,
studyID=studyID, trait=trait, threshold=threshold,
format=format, hyperlink=hyperlink)
write.csv2(stat,file=file.csv,row.names=FALSE)
}}
\usage{
getStudiesDB(GAlist = NULL)
}
\description{
#' @export
#'
writeStat <- function(GAlist=NULL, feature=NULL, featureID=NULL,
studyID=NULL, trait="T2D", threshold=1,
format="data.frame", file.csv=NULL, hyperlink=TRUE) {
stat <- getStat(GAlist=GAlist, feature=feature, featureID=featureID,
studyID=studyID, trait=trait, threshold=threshold,
format=format, hyperlink=hyperlink)
write.csv2(stat,file=file.csv,row.names=FALSE)
}
}
