% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_gact.R
\name{getVEGAS}
\alias{getVEGAS}
\title{Get VEGAS Data}
\usage{
getVEGAS(GAlist = NULL, ensg = NULL, studyID = NULL, rm.na = TRUE)
}
\arguments{
\item{GAlist}{A list object that contains directories and data required for the analysis, including the path to the Z-score data.}

\item{ensg}{Not currently used in the function, but can be reserved for future functionality.}

\item{studyID}{A character vector of study IDs to filter the Z-scores. Only Z-scores corresponding to these study IDs will be returned. If \code{NULL}, all studies are returned.}

\item{rm.na}{Logical; if \code{TRUE}, rows with \code{NA} values will be removed from the data. Defaults to \code{TRUE}.}
}
\value{
A matrix of Z-scores filtered by the provided \code{studyID} (if given) and cleaned of missing values if \code{rm.na} is \code{TRUE}.
}
\description{
This function retrieves VEGAS (Versatile Gene-based Association Study) Z-score data from the GAlist object, allowing optional filtering by \code{studyID} and handling of missing values. The function returns a filtered matrix of Z-scores corresponding to the specified \code{studyID} or all data if no specific study is provided.
}
\examples{
# Assuming GAlist is a valid object with the necessary structure
Z_scores <- getVEGAS(GAlist, studyID = c("study1", "study2"))

# Get all Z-scores and remove rows with NA values
Z_scores_all <- getVEGAS(GAlist, rm.na = TRUE)

}
