% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_gact.R
\name{getVEP}
\alias{getVEP}
\title{Retrieve VEP (Variant Effect Predictor) information for specified genes or rsids.}
\usage{
getVEP(GAlist = NULL, ensg = NULL, enst = NULL, ensp = NULL, rsids = NULL)
}
\arguments{
\item{GAlist}{A list object containing necessary data, including the directory (\code{dbdir}) where VEP files are located.}

\item{ensg}{A character vector of gene Ensembl IDs (e.g., ENSG00000139618) for which rsids should be retrieved and analyzed.}

\item{enst}{A character vector of transcript Ensembl IDs (optional, not used in the current function).}

\item{ensp}{A character vector of protein Ensembl IDs (optional, not used in the current function).}

\item{rsids}{A character vector of rsids (e.g., rs123) for which VEP annotations should be retrieved.}
}
\value{
A data frame containing the VEP annotations for the requested rsids, with columns:
\itemize{
\item \code{rsids}: The rsid (variant ID).
\item \code{Location}: The chromosomal location of the variant.
\item \code{REF_ALLELE}: The reference allele.
\item \code{Allele}: The alternate allele.
\item \code{Consequence}: The predicted consequence of the variant.
\item \code{IMPACT}: The predicted impact of the variant.
\item \code{SYMBOL}: The gene symbol.
\item \code{Gene}: The Ensembl gene ID.
\item \code{Feature_type}: The type of feature (e.g., transcript).
\item \code{Feature}: The Ensembl transcript ID.
\item \code{BIOTYPE}: The biotype of the transcript (e.g., protein_coding).
\item Additional columns for SIFT, PolyPhen, CLIN_SIG, and others.
}
}
\description{
This function fetches VEP annotations for either a list of genes (via \code{ensg}) or specific rsids.
It searches for the corresponding rsids if genes are provided and then extracts VEP data
from the Ensembl VEP files.
}
\details{
The function first checks whether rsids are provided directly or need to be retrieved
based on the provided Ensembl gene IDs (\code{ensg}). It then reads chromosome-specific
VEP files from the provided \code{GAlist} directory and extracts relevant variant annotations
for the specified rsids. Results are filtered and returned as a data frame.
}
\note{
The VEP files must be located in the \code{dbdir} of the \code{GAlist} object, and the files should be named
as \code{vep01.txt.gz}, \code{vep02.txt.gz}, ..., corresponding to each chromosome.
}
\examples{
# Example with Ensembl gene IDs
GAlist <- list(dbdir = "path_to_vep_files")
vep_data <- getVEP(GAlist, ensg = c("ENSG00000139618"))

# Example with rsids
vep_data <- getVEP(GAlist, rsids = c("rs123", "rs456"))

}
