% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gact.R
\name{gpath}
\alias{gpath}
\title{Generate gene sets for pathway analysis}
\usage{
gpath(
  GAlist = NULL,
  mcausal = 500,
  causal = NULL,
  overlap = FALSE,
  ngenes = 10,
  ncgenes = 2,
  nreps = 10,
  format = "markers"
)
}
\arguments{
\item{GAlist}{a list object containing SNP and pathway
data}

\item{mcausal}{an integer indicating the number of causal SNPs to sample.
Default is 500.}

\item{causal}{a vector of SNP IDs representing the causal SNPs to be used in
the gene set construction. Default is \code{NULL} and SNPs will be randomly
sampled from the available SNPs in \code{GAlist}.}

\item{ngenes}{an integer or vector indicating the number of genes to include
in each gene set. If a vector is supplied, a gene set will be generated for
each value in the vector. Default is 10.}

\item{ncgenes}{an integer or vector indicating the number of causal genes to
include in each gene set. If a vector is supplied, a gene set will be
generated for each combination of values in \code{ngenes} and \code{ncgenes}.
Default is 2.}

\item{nreps}{an integer indicating the number of gene sets to generate.
Default is 10.}
}
\value{
a list of gene sets, where each gene set is represented as a vector of
SNP IDs.
}
\description{
Generates gene sets for pathway analysis based on a given set of SNPs, using a
pathway database contained in a GAlist object. The function
samples causal SNPs and constructs gene sets of a given size and number of
causal genes.
}
\examples{
\dontrun{
# Generate gene sets
sets <- gpath(GAlist = GAlist, ngenes = 10, ncgenes = 2, nreps = 5)
}

}
