% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_gact.R
\name{simPathwayDB}
\alias{simPathwayDB}
\title{Simulate gene sets for pathway analysis}
\usage{
simPathwayDB(
  GAlist = NULL,
  mcausal = 500,
  causal = NULL,
  ngenes = 10,
  ncgenes = 2,
  nreps = 10,
  format = "markers"
)
}
\arguments{
\item{GAlist}{A list object containing SNP and pathway data.}

\item{mcausal}{Integer, number of causal SNPs to sample (default: 500).}

\item{causal}{Vector of SNP IDs for causal SNPs (default: NULL, SNPs are randomly sampled).}

\item{ngenes}{Integer or vector, number of genes per gene set (default: 10).}

\item{ncgenes}{Integer or vector, number of causal genes per gene set (default: 2).}

\item{nreps}{Integer, number of gene sets to generate (default: 10).}

\item{format}{String, format of the gene sets (default: "markers").}
}
\value{
List of gene sets, each represented as a vector of SNP IDs.
}
\description{
Generates gene sets for pathway analysis based on a given set of SNPs, using a
pathway database contained in a GAlist object. The function
samples causal SNPs and constructs gene sets of a given size and number of
causal genes.
}
\examples{
\dontrun{
  # Generate gene sets
  sets <- simPathwayDB(GAlist = GAlist, ngenes = 10, ncgenes = 2, nreps = 5)
}
}
