% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gact.R
\name{summaryDB}
\alias{summaryDB}
\title{Summarize GSEA and BLR results}
\usage{
summaryDB(GAlist = NULL)
}
\arguments{
\item{GAlist}{A list containing directory paths and potentially other information}
}
\value{
None
}
\description{
Summarize GSEA and BLR results
}
