% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gact.R
\name{updateStatDB}
\alias{updateStatDB}
\title{Update and process genetic association study summary statistics}
\usage{
updateStatDB(
  GAlist = NULL,
  stat = NULL,
  source = NULL,
  trait = "unknown",
  type = "unknown",
  gender = "unknown",
  ancestry = "unknown",
  build = "unknown",
  n = NULL,
  ncase = NULL,
  ncontrol = NULL,
  reference = "unknown",
  comments = "none",
  writeStatDB = TRUE,
  excludeMAFDIFF = 0.05
)
}
\arguments{
\item{GAlist}{A list object providing information and infrastructure of the gact database.}

\item{stat}{A dataframe of summary statistics for each SNP}

\item{source}{Source of the summary statistics}

\item{trait}{Trait being studied}

\item{type}{Type of trait (binary or quantitative)}

\item{gender}{Gender of study participants}

\item{ancestry}{Ancestry of study participants}

\item{build}{Genome build used in the study}

\item{n}{Sample size}

\item{ncase}{Number of cases}

\item{ncontrol}{Number of controls}

\item{reference}{Reference for the study}

\item{comments}{Additional comments}

\item{writeStatDB}{A flag indicating whether to perform quality control and write the processed summary statistics to a file}

\item{excludeMAFDIFF}{A threshold used to exclude SNPs with a difference in minor allele frequency (MAF) between cases and controls that exceeds this threshold}
}
\value{
A list of updated study information, including the study id, file name, trait, type, gender, ancestry, build, sample size, case/control count, reference, and comments.
}
\description{
This function updates and processes genetic association study summary statistics. The input to the function includes study information, summary statistics, and information about the sample.
}
