% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_gact.R
\name{addAnnotationDB}
\alias{addAnnotationDB}
\title{Add Annotation to Data Frame}
\usage{
addAnnotationDB(df = NULL, ensg = NULL, ensr = NULL, hyperlinkEXCEL = FALSE)
}
\arguments{
\item{df}{Data frame to which annotations are added.}

\item{ensg}{Vector of Ensembl gene IDs, if not provided, it's assumed to be rownames of df.}

\item{ensr}{Vector of Ensembl regulatory feature IDs, if not provided, it's assumed to be rownames of df.}

\item{hyperlinkEXCEL}{Boolean, if TRUE, adds hyperlinks to the Ensembl website for each gene ID.}
}
\value{
Annotated data frame.
}
\description{
This function adds annotations to a data frame based on Ensembl gene IDs. It can also create
hyperlinks for viewing in Excel.
}
