% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gact.R
\name{createDB}
\alias{createDB}
\title{createDB - Creates a database for genetic association studies}
\usage{
createDB(Glist = NULL, version = NULL, dbdir = NULL, what = "lite")
}
\arguments{
\item{Glist}{A list of genetic data. Default is NULL.}

\item{version}{Character string indicating the name of the database.
It is a required argument.}

\item{dbdir}{Character string indicating the directory where the database should be created.
Default is NULL.}

\item{what}{Character string indicating the type of database.
Default is "lite".}
}
\value{
A list containing the following items:
- version: Name of the database
- traits: NULL
- dirs: List of subdirectories within the database directory
- features: List of features in the database
- markers: Data frame of markers in the database
- rsids: List of rsids in the database
- cpra: List of cpra in the database
}
\description{
createDB - Creates a database for genetic association studies
}
\details{
The createDB function creates a database for genetic association studies.
If \code{dbdir} is not specified, the database will be created in the current working directory.
The function creates subdirectories within the database directory, including "glist", "gstat",
"gsets", "gsea", "ldsc", "gbayes", "marker", "download", and "drugdb".
The list \code{Glist} should contain the genetic data and will be filtered to only keep
markers that are present in \code{Glist$rsidsLD}. The filtered markers are stored in a data frame
and saved in the "marker" subdirectory as "markers.txt.gz".
}
\examples{
\dontrun{
createDB(Glist = genetic_data, version = "db1", dbdir = "/data/db1")
createDB(Glist = genetic_data, version = "db2", dbdir = "/data/db2", what = "full")
}

}
