% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gact.R
\name{gact}
\alias{gact}
\title{Create or download GACT data}
\usage{
gact(
  GAlist = NULL,
  version = NULL,
  task = "download",
  dbdir = NULL,
  what = "lite"
)
}
\arguments{
\item{GAlist}{A list object providing information and infrastructure of the gact database.}

\item{version}{The version of the gact database to use.}

\item{task}{The task to perform, either "download" or "process". By default "download".}

\item{dbdir}{The directory where the database should be stored, by default the current working directory.}

\item{what}{Specifies the type of data to download, either "lite" or "full". By default "lite".}
}
\value{
A list providing information and infrastructure of the gact database.
}
\description{
The gact function is used to create or download a GACT database with information
on genomic associations for complex traits. The genomic associations
include single marker associations linked to genetic markers in a Glist.
This includes stringent quality control. In addition genomic associations
linked to different genomic features (e.g. genes, proteins, chemical-complexes,
protein-complexes, biological pathways) are provided.
}
\examples{
\dontrun{
GAlist <- gact()
}
}
