% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_gact.R
\name{getMarkerSets}
\alias{getMarkerSets}
\title{Retrieve Marker Sets from Database}
\usage{
getMarkerSets(
  GAlist = NULL,
  feature = NULL,
  featureID = NULL,
  rsids = NULL,
  threshold = 0.01
)
}
\arguments{
\item{GAlist}{A list object with the gact database infrastructure.}

\item{feature}{A string specifying the biological feature type for the marker sets.}

\item{featureID}{ID of the feature to retrieve marker sets for.}

\item{rsids}{Optional vector of rsids to filter the marker sets.}

\item{threshold}{Threshold value for filtering (default is 0.01).}
}
\value{
A list of marker sets for the specified feature and feature ID, each set
being a character vector of rsids.
}
\description{
Retrieves marker sets based on a given feature and feature ID. Supports various features
such as Genes, Entrez Genes, Gene Symbol, Proteins, Pathways, GO, Protein Complexes,
Chemical Complexes, and others. Optionally, can filter by a list of SNP IDs.
}
\examples{
\dontrun{
  sets <- getMarkerSets(GAlist = GAlist, feature = "Genes",
                          featureID = c("ENSG00000165879", "ENSG00000012048"))
}
}
